(PC-17103) move tests files next to the tested component
