+++ src/features/bookOffer/components/BookingDetails.native.test.tsx
@@ -26,7 +26,7 @@ import {
 } from 'tests/utils'
 import { SnackBarHelperSettings } from 'ui/components/snackBar/types'
 
-import { BookingDetails } from '../BookingDetails'
+import { BookingDetails } from './BookingDetails'
 
 // eslint-disable-next-line local-rules/no-allow-console
 allowConsole({ error: true })
